Added custom hook for attendance componenet
