--- packages/attendance/src/stories/Loader.stories.tsx
@@ -1,7 +1,0 @@
-import * as React from "react";
-import Loader from "components/simple/Loader";
-import { storiesOf } from "@storybook/react";
-
-storiesOf("Loading", module).add("When loading", () => (
-  <Loader success="successful" fail="failure" />
-));
