+++ packages/attendance/src/utils/functions/ColorTheme.ts
@@ -1,0 +1,5 @@
+import colorTheme from "colorTheme";
+import { overrideColorTheme } from "@shiksha/common-lib";
+const colors = overrideColorTheme(colorTheme);
+
+export { colors, colorTheme };
