+++ packages/common-lib/src/services/attendanceRegistryService.js
@@ -90,7 +90,7 @@ export const update = async (data = {}, headers = {}) => {
   }
 }
 
-export const multipal = async (data = {}, header = {}) => {
+export const Multiple = async (data = {}, header = {}) => {
   let headers = {
     ...header,
     Authorization: 'Bearer ' + localStorage.getItem('token')
